Version 2.2.0.  * "faster" selection of locomotives, when selecting the loc address is set to white color.  * When a selected loc is controlled by another (XPressNet) device the loc symbol is turned yellow.
Tested with MDRRC-II (lite) and Roco LocMaus and in both situations the yellow symbol is not shown when expected.. On the Roco setup only when the set speed of the other device was zero, on MDRRC-II (Lite) somewhat vague.. Last one needs to be checked.

Be aware, settings will be lost!